Added CTF support with exported ELEKTA events (.eve) from Brainstorm (single event is exported). BIDS datasets are created for all subjects, with each RUN added as a separate SESSION.
Added jitter around events so the model doesn't learn a specific location of the annotation.
Added global min-max on each trial/NIFTI file.
Added downsampling parameter for the 3rd dimension - time (this is needed for model generalization on recordings with different Fs).

Note: The function MEG.py was modified to create BIDS datasets simulating a different subject for each session.